cabeza/expo_screen_time_pack/requirements.txt
ultralytics>=8.1.0
mediapipe==0.10.14
opencv-python>=4.9.0.80
numpy>=1.26.4
pymongo>=4.8.0
python-dotenv>=1.0.1
